Refactor performance script enqueue
